Updated code
Code with searching capabilities included into it
