+++ flask_learning/__init__.py
@@ -1,10 +1,8 @@
-import flask_sijax
 from flask import Flask
 from flask_sqlalchemy import SQLAlchemy
 from flask_bcrypt import Bcrypt
 from flask_login import LoginManager
 from flask_mail import Mail
-from flask_sijax import Sijax
 
 
 from flask_learning.config import Config
@@ -15,7 +13,6 @@ bcrypt = Bcrypt()
 login_manager = LoginManager()
 login_manager.login_view = 'users.user_login'
 login_manager.login_message_category = 'info'
-sijax = flask_sijax.Sijax()
 mail = Mail()
 
 
@@ -27,7 +24,6 @@ def create_app(config_class=Config):
     bcrypt.init_app(app)
     login_manager.init_app(app)
     mail.init_app(app)
-    sijax.init_app(app)
 
     from flask_learning.users.routes import users
     from flask_learning.posts.routes import posts
