+++ flask_learning/templates/navbar.html
@@ -1,28 +1,28 @@
 <!--<header class="site-header">-->
 <!--      <nav class="navbar navbar-expand-md navbar-dark bg-steel fixed-top">-->
 <!--        <div class="container">-->
-<!--          <a class="navbar-brand mr-4" href="{{ url_for('main.home_page') }}">Flask Blog</a>-->
+{#<!--          <a class="navbar-brand mr-4" href="{{ url_for('main.home_page') }}">Flask Blog</a>-->#}
 <!--          <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarToggle" aria-controls="navbarToggle" aria-expanded="false" aria-label="Toggle navigation">-->
 <!--            <span class="navbar-toggler-icon"></span>-->
 <!--          </button>-->
 <!--          <div class="collapse navbar-collapse" id="navbarToggle">-->
 <!--            <div class="navbar-nav me-auto">-->
-<!--              <a class="nav-item nav-link" href="{{ url_for('main.home_page') }}">Home</a>-->
-<!--              <a class="nav-item nav-link" href="{{ url_for('main.about_user') }}">About</a>-->
+{#<!--              <a class="nav-item nav-link" href="{{ url_for('main.home_page') }}">Home</a>-->#}
+{#<!--              <a class="nav-item nav-link" href="{{ url_for('main.about_user') }}">About</a>-->#}
 <!--            </div>-->
 <!--            &lt;!&ndash; Navbar Right Side &ndash;&gt;-->
-<!--            {% if current_user.is_authenticated %}-->
+{#<!--            {% if current_user.is_authenticated %}-->#}
 <!--            <div class="navbar-nav">-->
-<!--              <a class="nav-item nav-link" href="{{ url_for('posts.new_post') }}">Create Post</a>-->
-<!--              <a class="nav-item nav-link" href="{{ url_for('users.account') }}">Account</a>-->
-<!--              <a class="nav-item nav-link" href="{{ url_for('users.user_logout') }}">Logout</a>-->
-<!--            </div>-->
-<!--            {% else %}-->
-<!--            <div class="navbar-nav">-->
-<!--              <a class="nav-item nav-link" href="{{ url_for('users.user_login') }}">Login</a>-->
-<!--              <a class="nav-item nav-link" href="{{ url_for('users.user_register') }}">Register</a>-->
-<!--            </div>-->
-<!--            {% endif %}-->
+{#<!--              <a class="nav-item nav-link" href="{{ url_for('posts.new_post') }}">Create Post</a>-->#}
+{#<!--              <a class="nav-item nav-link" href="{{ url_for('users.account') }}">Account</a>-->#}
+{#<!--              <a class="nav-item nav-link" href="{{ url_for('users.user_logout') }}">Logout</a>-->#}
+{#<!--            </div>-->#}
+{#<!--            {% else %}-->#}
+{#<!--            <div class="navbar-nav">-->#}
+{#<!--              <a class="nav-item nav-link" href="{{ url_for('users.user_login') }}">Login</a>-->#}
+{#<!--              <a class="nav-item nav-link" href="{{ url_for('users.user_register') }}">Register</a>-->#}
+{#<!--            </div>-->#}
+{#<!--            {% endif %}-->#}
 <!--          </div>-->
 <!--        </div>-->
 <!--      </nav>-->
@@ -56,9 +56,9 @@
             </div>
             {% endif %}
         </ul>
-        <form class="d-flex" method="post" action="{{ url_for('posts.search_post')}}">
-          {{form.hidden_tag()}}
-          <input class="form-control me-2" type="search" placeholder="Search" aria-label="Search", name="searched">
+        <form class="d-flex" method="post" action="{{ url_for('main.search_post') }}">
+            {{ form.hidden_tag() }}
+          <input class="form-control me-2" type="search",  name="searched" placeholder="Search" aria-label="Search"/>
           <button class="btn btn-outline-success" type="submit">Search</button>
         </form>
       </div>
