+++ flask_learning/templates/search.html
@@ -1,8 +1,7 @@
 {% extends 'layout.html' %}
 
 {% block content %}
-
-<!--    {{post_searched}}-->
+{#    {{searched}}#}
 
     {% for post in posts %}
         <article class="d-flex content-section">
@@ -10,12 +9,23 @@
           <div class="media-body">
             <div class="article-metadata">
               <a class="me-2" href="{{url_for('users.user_posts', username=post.author.username) }}">{{ post.author.username }}</a>
-              <small class="text-muted">{{ post.date_posted.strftime('%Y-%m-%d ') }}</small>
+              <small class="text-muted">{{ post.date_posted.strftime('%Y-%m-%d') }}</small>
             </div>
             <h2><a class="article-title" href="{{ url_for('posts.post', post_id=post.id) }}">{{ post.title }}</a></h2>
             <p class="article-content">{{ post.content }}</p>
           </div>
         </article>
+    {% endfor %}
+
+    {% for page_num in posts.iter_pages(left_edge=1, right_edge=1, left_current=1, right_current=2) %}
+        {% if page_num %}
+            {% if posts.page == page_num %}
+                <a class="btn btn-primary mb-4" href="{{ url_for('main.home_page', page=page_num) }}">{{ page_num }}</a>
+            {% else %}
+                <a class="btn btn-outline-info mb-4" href="{{ url_for('main.home_page', page=page_num) }}">{{ page_num }}</a>
+            {% endif %}
+        {% else %}
+        {% endif %}
     {% endfor %}
 
 {% endblock content %}
